Cloud Deployment/requirements.txt
# Function dependencies, for example:
# package>=version

tensorflow>=2.9.1
google-cloud-storage
Pillow
numpy
flask